requirements.txt
PyQt6==6.10.0
reportlab==4.4.5
python-barcode==0.16.0
pillow==11.0.0